[softeer] 성적평가
